Use explicit observability pairings in Studio
