+++ ui/hooks/use-stream-events.test.tsx
@@ -301,6 +301,7 @@ describe("useStreamEvents", () => {
         expiresAt: null,
         name: "prisma-wal",
         nextOffset: "3",
+        observability: null,
         profile: null,
         sealedThrough: "-1",
         uploadedThrough: "-1",
@@ -373,6 +374,7 @@ describe("useStreamEvents", () => {
         expiresAt: null,
         name: "golden-stream-2",
         nextOffset: "2",
+        observability: null,
         profile: null,
         sealedThrough: "-1",
         uploadedThrough: "-1",
@@ -448,6 +450,7 @@ describe("useStreamEvents", () => {
         expiresAt: null,
         name: "golden-stream-2",
         nextOffset: "200",
+        observability: null,
         profile: null,
         sealedThrough: "-1",
         uploadedThrough: "-1",
@@ -499,6 +502,7 @@ describe("useStreamEvents", () => {
         expiresAt: null,
         name: "gharchive-demo-all",
         nextOffset: "1",
+        observability: null,
         profile: null,
         sealedThrough: "-1",
         uploadedThrough: "-1",
@@ -547,6 +551,7 @@ describe("useStreamEvents", () => {
         expiresAt: null,
         name: "gharchive-demo-all",
         nextOffset: "1",
+        observability: null,
         profile: null,
         sealedThrough: "-1",
         uploadedThrough: "-1",
@@ -599,6 +604,7 @@ describe("useStreamEvents", () => {
       expiresAt: null,
       name: "prisma-wal",
       nextOffset: "100",
+      observability: null,
       profile: null,
       sealedThrough: "-1",
       uploadedThrough: "-1",
@@ -672,6 +678,7 @@ describe("useStreamEvents", () => {
         expiresAt: null,
         name: "prisma-wal",
         nextOffset: "120",
+        observability: null,
         profile: null,
         sealedThrough: "-1",
         uploadedThrough: "-1",
@@ -792,6 +799,7 @@ describe("useStreamEvents", () => {
         expiresAt: null,
         name: "prisma-wal",
         nextOffset: "5",
+        observability: null,
         profile: null,
         sealedThrough: "-1",
         uploadedThrough: "-1",
@@ -900,6 +908,7 @@ describe("useStreamEvents", () => {
         expiresAt: null,
         name: "prisma-wal",
         nextOffset: "6",
+        observability: null,
         profile: null,
         sealedThrough: "-1",
         uploadedThrough: "-1",
@@ -952,6 +961,7 @@ describe("useStreamEvents", () => {
         expiresAt: null,
         name: "prisma-wal",
         nextOffset: "6",
+        observability: null,
         profile: null,
         sealedThrough: "-1",
         uploadedThrough: "-1",
@@ -1024,6 +1034,7 @@ describe("useStreamEvents", () => {
       expiresAt: null,
       name: "prisma-wal",
       nextOffset: "3",
+      observability: null,
       profile: null,
       sealedThrough: "-1",
       uploadedThrough: "-1",
